TH helper to fix row types when ready to run.
This is an experiment with what I fear is disastrously low
usability. Might be worth thinking about for a little while before
rolling it back.

The idea is this:

When writing a function that might return an error, we don't want to
pin down the error type, so we say,

```
fun1 :: TooBig ∈ rs => Int -> Either (CoRec Identity rs) ()
```

To indicate that this function might return `TooBig`, so the sum type
must have that as a variant.

Each function can add its own constraints, leaving you with a
situation where you want to finally apply some handlers to an actual
error. At this point, you are dealing with something like,

```
forall rs. (TooBig ∈ rs, Even ∈ rs) => CoRec Identity rs
```

And we want to say that `rs ~ '[TooBig, Even]`. Except we don't want
to have to say it, we want the composite constraint to fix the type of
`rs` to the union of the constraints.

What I've done here is added some Template Haskell to do something
like this. The problem is that the template haskell must reify a Name,
and that Name must be declared in a previous declaration group. This
means that the entire expression whose type we want to fix must be
declared at the top level, and we likely need a seemingly extraneous
"return []" declaration splice to create a new declaration group. Then
we can pass the Name to the template haskell that can generate a type
annotation derived from the constraints.